requirements.txt
openai==1.13.3
azure-mgmt-resource==23.0.1
azure-mgmt-search==9.1.0
azure-mgmt-cognitiveservices==13.5.0
azure-identity==1.16.0
azure-search-documents==11.4.0
azure-ai-ml==1.16.0
promptflow[azure]==1.11.0
promptflow-tracing==1.11.0
# promptflow-tools==1.4.0
# promptflow-rag==0.1.0
# promptflow-evals==0.3.0
python-dotenv

# The following dependencies are required for provisioning
# azure dependencies
azure-core==1.30.1
pyperclip


# utilities
streamlit